# docs/swagger.yaml
definitions:
...
  models.CreateUserInfo:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      realm:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - realm
    - username
    type: object
  models.CredentialRepresentation:
    properties:
      createdDate:
        type: integer
      credentialData:
        type: string
      id:
        type: string
      type:
        type: string
      userLabel:
        type: string
    required:
    - createdDate
    - credentialData
    - id
    - type
    type: object
...
  models.GetUserInfo:
    properties:
      Account:
        type: string
      AccountId:
        type: string
      OpenId:
        type: string
      attributes:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      createDate:
        type: string
      createdTimestamp:
        type: integer
      creator:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      firstName:
        type: string
      groups:
        type: string
      id:
        type: string
      lastName:
        type: string
      modifier:
        type: string
      modifyDate:
        type: string
      phoneNumber:
        type: string
      realm:
        type: string
      requiredActions:
        items:
          type: string
        type: array
      roles:
        type: string
      username:
        type: string
    type: object
  models.GroupData:
    properties:
      id:
        type: string
      name:
        type: string
      path:
        type: string
    required:
    - id
    - name
    - path
    type: object
...
  models.Id:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  models.ResetUserPasswordInfo:
    properties:
      password:
        type: string
      passwordConfirm:
        type: string
      temporary:
        type: boolean
    required:
    - password
    - passwordConfirm
    type: object
...
  models.UserIdProviderData:
    properties:
      identityProvider:
        type: string
      userId:
        type: string
      userName:
        type: string
    required:
    - identityProvider
    - userId
    - userName
    type: object
info:
  contact: {}
  title: IAM.Backend API Document
  version: "1.0"
paths:
  /account/{accountId}/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetUserInfo'
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 목록
      tags:
      - Account
    post:
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Id'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 생성
      tags:
      - Account
  /account/{accountId}/users/{userId}:
    get:
      parameters:
      - description: account Id
        in: path
        name: accountId
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserInfo'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 상세정보 조회
      tags:
      - Account
    put:
      parameters:
      - description: account Id
        in: path
        name: accountId
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserInfo'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 정보 변경
      tags:
      - Account
  /account/{accountId}/users/{userId}/credentials:
    get:
      parameters:
      - description: account Id
        in: path
        name: accountId
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CredentialRepresentation'
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 자격증명 조회
      tags:
      - Account
  /account/{accountId}/users/{userId}/federated-identity:
    get:
      parameters:
      - description: account Id
        in: path
        name: accountId
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserIdProviderData'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 ID 제공자 조회
      tags:
      - Account
  /account/{accountId}/users/{userId}/federated-identity/{providerId}:
    delete:
      parameters:
      - description: account Id
        in: path
        name: accountId
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: Provider Id
        in: path
        name: providerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 ID 제공자 제거
      tags:
      - Account
  /account/{accountId}/users/{userId}/reset-password:
    put:
      parameters:
      - description: account Id
        in: path
        name: accountId
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: ResetUserPasswordInfo
        in: body
        name: roleId
        required: true
        schema:
          $ref: '#/definitions/models.ResetUserPasswordInfo'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 유저 비밀번호 변경
      tags:
      - Account
  /account/{accountId}/users/me:
    put:
      parameters:
      - description: account Id
        in: path
        name: accountId
        required: true
        type: string
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserInfo'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Account 자신의 계정 정보 변경
      tags:
      - Account
...
  /users/{userId}:
    delete:
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 삭제
      tags:
      - Users
  /users/{userId}/groups:
    delete:
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: Group Id
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 그룹 탈퇴
      tags:
      - Users
    get:
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: data start count
        in: query
        name: first
        required: true
        type: integer
      - description: data max count
        in: query
        name: max
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GroupData'
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 그룹 조회
      tags:
      - Users
    put:
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: Group Id
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 그룹 가입
      tags:
      - Users
  /users/{userId}/logout:
    post:
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 전체 세션 제거
      tags:
      - Users
  /users/{userId}/sessions:
    get:
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SesstionData'
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 세션 조회
      tags:
      - Users
  /users/{userId}/sessions/{sessionId}:
    delete:
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: Session Id
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SesstionData'
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 세션 제거
      tags:
      - Users
  /users/initialize:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Users 유저 초기 설정 작업(대상이 자기 자신인 경우에만)
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"